refactor(web): consolidate duplicated SettingRow into a shared component
